Mudando double pra int

+++ Trabalho Final FTS/Calculadora.cs
@@ -8,19 +8,19 @@ namespace Trabalho_Final_FTS
 {
     public class Calculadora
     {
-        public static double Somar(double a, double b)
+        public static int Somar(int a, int b)
         {
             return 0;
         }
-        public static double Subtrair(double a, double b)
+        public static int Subtrair(int a, int b)
         {
             return 0;
         }
-        public static double Multiplicar(double a, double b)
+        public static int Multiplicar(int a, int b)
         {
             return 0;
         }
-        public static double Dividir(double a, double b)
+        public static int Dividir(int a, int b)
         {
             return (a / b);
         }
